backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-jose
python-multipart
gunicorn
pandas
python-dotenv
scikit-learn
prophet
